delete container folder and move footer.js to footer folder
