+++ React/src/components/footer/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Container, Wrapper, Row, Column, Link, Title } from'./styles/footer'
+import { Container, Wrapper, Row, Column, Link, Title } from'./footer_style'
 
 export default function Footer({ children, ...restProps}){
     return <Container{...restProps}>{children}</Container>
