+++ React/src/index.js
@@ -2,10 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
 import './index.css';
-import { GlobalStyles } from './global-styles'
 import App from './App';
 import Navbar from "./components/Navbar/Navbar"
-import { FooterContainer } from "./containers/footer"
+import { FooterContainer } from "./components/footer/footer"
 
 
 ReactDOM.render(
